requirements.txt
alembic==1.7.7
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blis==0.7.7
Boruta==0.3
boto3==1.21.28
botocore==1.24.28
Bottleneck==1.3.4
brotlipy==0.7.0
catalogue==1.0.0
certifi==2020.6.20
cfbd==4.3.2
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.0.0
colorlover==0.3.0
cryptography==36.0.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
databricks-cli==0.16.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
docker==5.0.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
fonttools==4.33.3
funcy==1.17
future==0.18.2
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.7.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
kmodes==0.12.1
lightgbm==3.3.2
llvmlite==0.37.0
Mako==1.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mlflow==1.25.1
mlxtend==0.19.0
multimethod==1.8
murmurhash==1.0.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
nltk==3.7
notebook==6.4.11
numba==0.54.1
numexpr==2.8.1
numpy==1.20.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.0
pip==21.2.4
plac==1.1.3
plotly==5.7.0
preshed==3.0.6
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycaret==2.3.10
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
PyJWT==2.3.0
pyLDAvis==3.2.2
pynndescent==0.5.6
pyod==1.0.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==22.3.0
querystring-parser==1.2.4
regex==2022.4.24
requests==2.27.1
s3transfer==0.5.2
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==61.2.0
six==1.16.0
smart-open==6.0.0
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==2.3.7
SQLAlchemy==1.4.36
sqlparse==0.4.1
srsly==1.0.5
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado==0.13.3
textblob==0.17.1
thinc==7.4.5
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
umap-learn==0.5.3
urllib3==1.26.9
visions==0.7.4
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.6.0
wordcloud==1.8.1
yellowbrick==1.3.post1
zipp==3.8.0
